requirements.txt
librosa
pyserial
numpy
tqdm
